rename Nav component
