+++ src/helm-frontend/src/components/NavBar/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./NavBar";
